Added notification type and finished notifications

+++ DieteticSNS/DieteticSNS.Application/Models/Followings/Commands/UnfollowUser/UnfollowUserCommandHandler.cs
@@ -3,7 +3,9 @@ using System.Threading;
 using System.Threading.Tasks;
 using DieteticSNS.Application.Common.Exceptions;
 using DieteticSNS.Application.Common.Interfaces;
+using DieteticSNS.Application.Models.Notifications.Commands.CreateNotification;
 using DieteticSNS.Domain.Entities;
+using DieteticSNS.Domain.Enumerations;
 using MediatR;
 
 namespace DieteticSNS.Application.Models.Followings.Commands.UnfollowUser
@@ -11,10 +13,12 @@ namespace DieteticSNS.Application.Models.Followings.Commands.UnfollowUser
     public class UnfollowUserCommandHandler : IRequestHandler<UnfollowUserCommand>
     {
         private readonly IDieteticSNSDbContext _context;
+        private readonly IMediator _mediator;
 
-        public UnfollowUserCommandHandler(IDieteticSNSDbContext context)
+        public UnfollowUserCommandHandler(IDieteticSNSDbContext context, IMediator mediator)
         {
             _context = context;
+            _mediator = mediator;
         }
 
         public async Task<Unit> Handle(UnfollowUserCommand request, CancellationToken cancellationToken)
@@ -30,7 +34,28 @@ namespace DieteticSNS.Application.Models.Followings.Commands.UnfollowUser
 
             _context.Followings.Remove(entity);
 
-            await _context.SaveChangesAsync(cancellationToken);
+            if (await _context.SaveChangesAsync(cancellationToken) > 0)
+            {
+                var setting = _context.UserNotificationSettings
+                        .Where(x => x.UserId == entity.FollowerId)
+                        .FirstOrDefault()
+                        .UserFollowings;
+
+                if (setting)
+                {
+                    if (entity.FollowerId != entity.UserId)
+                    {
+                        var command = new CreateNotificationCommand
+                        {
+                            UserId = entity.FollowerId,
+                            RecipientId = entity.UserId,
+                            NotificationType = NotificationType.UserUnfollowing
+                        };
+
+                        await _mediator.Send(command);
+                    }
+                }
+            }
 
             return Unit.Value;
         }
